Style/Rewrite Instance class
* use class member initialization
* move db load/delete/update stuff to InstanceMgr class
* split code to seperate file

Speaking of changes, the design was pretty weird. The legacy code changes the database without updating the saved data. Over all we have the class InstanceMgr which should take care about the data management.
There is still a lot to do in InstanceMgr since it is not thread save, we should use modern singleton implementation to get rid of some logic problems.

+++ src/world/Chat/Commands/InstanceCommands.cpp
@@ -141,7 +141,7 @@ bool ChatHandler::HandleGetInstanceInfoCommand(const char* args, WorldSession* m
 
         ss << "\n";
     }
-    ss << "Created: " << MSG_COLOR_CYAN << Util::GetDateTimeStringFromTimeStamp((uint32)instance->m_creation) << "|r\n";
+    ss << "Created: " << MSG_COLOR_CYAN << Util::GetDateTimeStringFromTimeStamp(instance->m_creation) << "|r\n";
     if (instance->m_expiration != 0)
         ss << "Expires: " << MSG_COLOR_CYAN << Util::GetDateTimeStringFromTimeStamp((uint32)instance->m_expiration) << "|r\n";
